Add Collection Center screen
Essentially the same as Camp Management screen except there is no
button to view list of inhabitants. Complete with a function to set
Metadata. This currently shares the same modal as Camp screen metadata
because they are essentially the same.

Also rename the CampMgmtScreenStyle to CampAndCollectionScreenStyle
since this is used for both screens.